@@ -4,13 +4,13 @@ namespace Laravel\Cashier\Tests;
 
 use Carbon\Carbon;
 use Illuminate\Http\Request;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Laravel\Cashier\Tests\Fixtures\User;
 use Illuminate\Database\Capsule\Manager as DB;
 use Illuminate\Database\Eloquent\Model as Eloquent;
 use Laravel\Cashier\Tests\Fixtures\CashierTestControllerStub;
 
-class CashierTest extends PHPUnit_Framework_TestCase
+class CashierTest extends TestCase
 {
     public static function setUpBeforeClass()
     {
