@@ -3,10 +3,10 @@
 namespace Laravel\Cashier\Tests;
 
 use Illuminate\Http\Request;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Laravel\Cashier\Tests\Fixtures\WebhookControllerTestStub;
 
-class WebhookControllerTest extends PHPUnit_Framework_TestCase
+class WebhookControllerTest extends TestCase
 {
     public function testProperMethodsAreCalledBasedOnStripeEvent()
     {
